README: remove daemon mode from roadmap
I already have a basically working daemon in the local daemon-prototype
branch, but I realized it's a bad idea: it's not atypical to need
multiple instances generating feeds for multiple users. Adding one or
several cron jobs is really simple and painless for anyone worth their
salt, so this idea is canned.